[remote] -Fix: double arg parser should return double, not int
git-svn-id: svn://repo.hu/pcb-rnd/trunk@5136 1d61de99-560f-4dba-8957-1e505d42358b

+++ src_plugins/hid_remote/proto_lowparse.h
@@ -297,7 +297,7 @@ static int aint(proto_node_t *nd, int badval)
 	return l;
 }
 
-static int adbl(proto_node_t *nd, double badval)
+static double adbl(proto_node_t *nd, double badval)
 {
 	char *s = astr(nd), *end;
 	double d;
